Swap out the net/rpc with gRPC (#6)
* Add basic grpc and protobuf setup

* Copy over all the types into a proto representation

* Add connect/close

* Update all methods to use the proto version instead

* Update tests and linting

* Don't install go using brew

* Update client and server to use grpc

* lint

+++ pkg/znode/znode.go
@@ -10,7 +10,7 @@ const (
 type ZNode struct {
 	// ZNode metadata.
 	Name               string
-	Version            int
+	Version            int64
 	Children           map[string]*ZNode
 	NodeType           ZNodeType
 	NextSequentialNode int
@@ -19,7 +19,7 @@ type ZNode struct {
 	Data []byte
 }
 
-func NewZNode(name string, version int, nodeType ZNodeType, data []byte) *ZNode {
+func NewZNode(name string, version int64, nodeType ZNodeType, data []byte) *ZNode {
 	return &ZNode{
 		Name:    name,
 		Version: version,
